Remove binary characters that crash node
@yozlet was having some issues adding one of his
repositories to federalist and it turns out that
some of the output from the build process was colored,
and those color characters were causing the JSON
parser to crash in node. This resulted in bad
communication between the builder and the main
application.

Now output should be scrubbed using `tr`