Add callback to federalist builder
This commit adds a callback to Federalist Builder so the container can let Federalist Builder know that it has finished building and is available.

This commit also tells the container to sleep after it has finished running so that it will not try to restart and run the build again.