Only set AWS_DEFAULT_REGION if it is not already set
The `AWS_DEFAULT_REGION` is the variable that describes which region the site will be published to.

Before this commit, the region was hardcoded to "us-east-1".
This created problems migrating to GovCloud.

Requiring the AWS region to be explicitly set invites its own problems because it requires us to modify the production app running in E/W.

This commit configures the build environment so the AWS region can be overriden, but defaults to "us-east-1".
This allows us to support legacy behavior, while being able to provide an override to publish sites to S3 in GovCloud.

closes #19