Initialize nvm before build script
This commit reverts changes in 280b15961363cbae4dbe3b39f2016e08ccdd96eb
and initializes nvm before using it in the build script.

Even though using bash in login mode makes nvm available in login mode to the
script it is invoked on, it does not make it available to child scripts
that script invokes. Because of this, login mode cannot be used to make
nvm available to build.sh. This change removes login mode and
initializes nvm in build.sh before using it.