Remove git details from site config
Prior to this commit, details about the commit were added to the site's
config before the site was built. This created problems when the
commit's message had syntax that resulted in malformed yaml, for
example, a `"` character in a commit message. This would create yaml
parse errors when the site was building, failing the build.

This commit removes that commit data since it does not seem to be used
by any sites, and there doesn't seem to be a straightforward fix for the
issue above:

Ref #823