Stop scary "missing federalist script" warning
When a build a `pacakge.json` file, the build container would run `npm run federalist || true`. This allowed users to add custom npm scripts, but had a few problems:

- Sites without a `federalist` script would see a disconcerting npm error about a missing federalist script
- Sites with a broken `federalist` script would not see their builds fail

This commit adds a conditional that checks whether the script is present in the `package.json` before attempting to run it. This means that the error will no longer appear for sites w/o a federalist script. In addition, the build will fail if the federalist script has a nonzero exit code.

Ref #33