Improve build container logging and remove superfluous activities
This commit is the result of work to debug an issue that was noticed on staging where federalist-builder was reporting that all builds were timing out for some larger sites. The issue was very hard to discover due to the lack of visibility into the container's activities.

This commit adds additional visibility, and removes several superfluous activities that were discovered to be the cause of the mysterious build timeouts. The changes are:

- Use `tee` to follow logs from build scripts in real time in container logs
- Add logging statements to make it clear where in the process the build is
- Use a base64 encoded string for the output in the build logs callback to avoid weird javascript string parsing errors
- Limit the length of build log output when calling back to the log callback to keep excessively long logs from exceeding bash argument length
- Preserve newlines in callbacks back to Federalist
- Remove redirect creation in `publish.sh` since it is long-running and unnecessary in an S3 bucket with a propper website configuration