Allow sites to use RVM to swtich ruby versions
This commit adds the ability for sites to set their ruby version using a `.ruby-version` file.

This commit adds a Federalist user to build the sites since running RVM as root can be kind of shaky. It also moves more of the configuration for the ruby install into build.sh. This is necessary because switching ruby versions means the old configurations will be left behind with the old version.

Ref #6