Unset pipefail before using pipes that can fail
We have several points where we do something like this:

```shell
curl some.url.gov ; curl some.other.url.gov || true
```

The purpose of this is to not break the build if any external services are down. This is very useful when reporting build logs, since the request is sent in the middle of the build process.

Additonally, we have build steps where we use `tee` to output the build output into the logs. That looks like this:

```shell
/app/build.sh 2>&1 | tee /dev/stderr
```

When `/app/build.sh` fails there, we want the entire build to fail.

The result of the above is that we want want the container to have `pipefail` enable some of the time, and disabled other times. This commit sets it when the build starts, and momentarily disables it when we do not want it.