Fix issue where timed out builds reported success.
This commit fixes an issue that arose with builds that timed out. When builds would time out, cloud.gov would stop the build would would trigger a call to the `post`. Since the exit code was zero, this would report to Federalist that the build completed successfully, even though it timed out.

This commit changes the `main.sh` script to detect when a build does not complete before exiting, and reports an error in that case, with an error message hinting that the build may have timed out.