Timeout builds after they run for 20 minutes
This commit adds a new script named `run.sh` which invokes `main.sh`
with an instruction to time it out after 20 minutes. This means that if
the process building the site takes longer than 20 minutes to execute,
this new script will stop it.

After setting the timeout, the run script then sleeps, and invocation
that was removed from the main script.

The introduction of a new script was necessary because `timeout` cannot
be used to run a function. Complex commands must be contained inside a
separate script and run that way.