Only publish files that have changed since the last publish
This commit replaces `publish.sh` with a slightly more sophisticated ruby script in `publish.rb`. This script downloads the MD5 hashes of the objects in the buckets and compares them to the MD5 hashes of the local files before uploading. It uses this information to determine if a file has changed since the files were last published, and only re-uploads files that have a different checksum.

This commit also create redirect objects for pages with a path in the form of `/path/to/page/`, so that `/path/to/page` will redirect to the correct place concerning the base url.

Ref #927